% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{exprs}
\alias{exprs}
\alias{exprs<-}
\alias{ret_classes}
\alias{ret_classes<-}
\alias{hashes}
\alias{hashes<-}
\alias{hData}
\alias{hData<-}
\title{Histry object accessors}
\usage{
exprs(obj)

exprs(obj) <- value

ret_classes(obj)

ret_classes(obj) <- value

hashes(obj)

hashes(obj) <- value

hData(obj)

hData(obj) <- value
}
\arguments{
\item{obj}{The object to access components of.}
}
\description{
Getters and setters for histry-related objects
}
