% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveload.R
\docType{methods}
\name{combineHistry}
\alias{combineHistry}
\alias{combineHistry,HistoryData-method}
\alias{combineHistry,VirtHistoryTracker-method}
\alias{combineHistry,HistoryData-method}
\title{Combine History data from two sources/records}
\usage{
combineHistry(x, y, before = FALSE)

\S4method{combineHistry}{HistoryData}(x, y, before = FALSE)

\S4method{combineHistry}{VirtHistoryTracker}(x, y, before = FALSE)
}
\arguments{
\item{x}{An object representing captured history}

\item{y}{An object representing captured history}

\item{before}{logical. Should the data from y be placed before the
data from x in the combined history. Defaults to \code{FALSE}}
}
\value{
The same class of object as \code{x}. I.e., if x is a
    HistoryTracker object, a HistoryTracker (of the same subclass)
    is returned. If X is a \code{HistoryData} object, that class
    of object is returned.
}
\description{
Combines the underlying history data from two
    representations.
}
\note{
HistoryTracker objects are reference classes, meaning that
    if \code{x} is a HistoryTracker the changes will be reflected
    in all variables representing that trackr.
}
