% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{evaltracer}
\alias{evaltracer}
\title{evaltracer}
\usage{
evaltracer(on = TRUE)
}
\arguments{
\item{on}{logical. Are we turning evaluation tracing on or off?}
}
\description{
This function is exported due to vagaries of how tracing works. It should not
ever be called directly by the end user
}

