% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking.R
\docType{data}
\name{histropts}
\alias{histropts}
\title{histry session options environment
Environment where the default history tracker is set
Users should not directly set or retrieve things from this environemnt,
it is exported for use by packages that use histry.}
\format{An object of class \code{environment} of length 1.}
\usage{
histropts
}
\keyword{datasets}

