% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prov.R
\name{histryProvDF}
\alias{histryProvDF}
\title{Generate ProvStoreDF from histry}
\usage{
histryProvDF(tracker = histry_tracker())
}
\arguments{
\item{tracker}{The HistoryTracker object to mine for the prov table information.}
}
\value{
a ProvStoreDF object
}
\description{
Generates a tabular 'provenance store' from the history
of evaluated expressions captured by histry
}
