% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{knitrtracer}
\alias{hist_within_doc}
\alias{knitrtracer}
\title{knitr history tracking

These functions turn on history tracking within the knitr weaving
process. \code{knitrtracer} does so from the R session, whereas
\code{hist_within_doc} is meant to be called within the first
code block of a knitr document.}
\usage{
knitrtracer(on)

hist_within_doc()
}
\arguments{
\item{on}{logical. Should tracking be turned on (TRUE) or off (FALSE)}
}

