% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{knitrtracer}
\alias{knitrtracer}
\alias{evaltracer}
\title{knitr history tracking

These functions are exported due to the vagaries of how tracing
functions works in R. Knitr history support is now turned on when
the histry package is loaded; They should never be called directly
by an end user.}
\usage{
knitrtracer(on, record = FALSE)

evaltracer(on = TRUE, record = FALSE)
}
\arguments{
\item{on}{logical. Should tracking be turned on (TRUE) or off (FALSE)}

\item{record}{logical. Should visibly printed results within the weaving
process be recorded (if trackr is available).}
}
\description{
evaltracer
}
