% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{historyTracker}
\alias{historyTracker}
\alias{knitrTracker}
\alias{knitr_tracker}
\title{Construct and activate a HistoryTracker object}
\usage{
historyTracker(id = "history_tracker")

knitrTracker()
}
\arguments{
\item{id}{character. The id of the taskback}
}
\value{
A HistoryTracker object
}
\description{
Constructors for the HistoryTrackr family of objects.
}
\note{
Generally end-users should never need to call these functions directly. Simply
loading the histry package should provide them with history tracking.
}
