% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{methods}
\name{exprs}
\alias{exprs}
\alias{exprs,HistoryData-method}
\alias{exprs,HistoryData}
\alias{exprs,VirtHistoryTracker-method}
\alias{exprs,VirtHistoryTracker}
\alias{exprs<-}
\alias{exprs<-,HistoryData-method}
\alias{exprs<-,HistoryData}
\alias{exprs<-,VirtHistoryTracker-method}
\alias{exprs<-,VirtHistoryTracker}
\alias{ret_classes}
\alias{ret_classes,HistoryData-method}
\alias{ret_classes,HistoryData}
\alias{ret_classes,VirtHistoryTracker-method}
\alias{ret_classes,VirtHistoryTracker}
\alias{ret_classes<-}
\alias{ret_classes<-,HistoryData-method}
\alias{ret_classes<-,HistoryData}
\alias{ret_classes<-,VirtHistoryTracker-method}
\alias{ret_classes<-,VirtHistoryTracker}
\alias{hashes}
\alias{hashes,HistoryData-method}
\alias{hashes,HistoryData}
\alias{hashes,VirtHistoryTracker-method}
\alias{hashes,VirtHistoryTracker}
\alias{hashes<-}
\alias{hashes<-,HistoryData-method}
\alias{hashes<-,HistoryData}
\alias{hashes<-,VirtHistoryTracker-method}
\alias{hashes<-,VirtHistoryTracker}
\alias{hData}
\alias{hData,HistoryData-method}
\alias{hData,HistoryData}
\alias{hData,VirtHistoryTracker-method}
\alias{hData,VirtHistoryTracker}
\alias{hData<-}
\alias{hData<-,HistoryData-method}
\alias{hData<-,HistoryData}
\alias{hData<-,VirtHistoryTracker-method}
\alias{hData<-,VirtHistoryTracker}
\title{Histry object accessors}
\usage{
exprs(obj)

\S4method{exprs}{HistoryData}(obj)

\S4method{exprs}{VirtHistoryTracker}(obj)

exprs(obj) <- value

\S4method{exprs}{HistoryData}(obj) <- value

\S4method{exprs}{VirtHistoryTracker}(obj) <- value

ret_classes(obj)

\S4method{ret_classes}{HistoryData}(obj)

\S4method{ret_classes}{VirtHistoryTracker}(obj)

ret_classes(obj) <- value

\S4method{ret_classes}{HistoryData}(obj) <- value

\S4method{ret_classes}{VirtHistoryTracker}(obj) <- value

hashes(obj)

\S4method{hashes}{HistoryData}(obj)

\S4method{hashes}{VirtHistoryTracker}(obj)

hashes(obj) <- value

\S4method{hashes}{HistoryData}(obj) <- value

\S4method{hashes}{VirtHistoryTracker}(obj) <- value

hData(obj)

\S4method{hData}{HistoryData}(obj)

\S4method{hData}{VirtHistoryTracker}(obj)

hData(obj) <- value

\S4method{hData}{HistoryData}(obj) <- value

\S4method{hData}{VirtHistoryTracker}(obj) <- value
}
\arguments{
\item{obj}{The object to access components of.}

\item{value}{The new value for the field.}
}
\description{
Getters and setters for histry-related objects
}
